Re-organize code and added comments to match `graphql-js`.
