+++ types/scalars.go
@@ -6,104 +6,6 @@ import (
 	"strconv"
 )
 
-type SerializeFn func(value interface{}) interface{}
-type ParseValueFn func(value interface{}) interface{}
-type ParseLiteralFn func(valueAST ast.Value) interface{}
-type GraphQLScalarTypeConfig struct {
-	Name         string `json:"name"`
-	Description  string `json:"description"`
-	Serialize    SerializeFn
-	ParseValue   ParseValueFn
-	ParseLiteral ParseLiteralFn
-}
-
-// GraphQLScalarType implements GraphQLType, GraphQLInputType, GraphQLNamedType,
-// 								GraphQLOutputType, etc
-// (TODO: find out what other interfaces GraphQLScalarType implements)
-type GraphQLScalarType struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-
-	scalarConfig GraphQLScalarTypeConfig
-
-	err error
-}
-
-func NewGraphQLScalarType(config GraphQLScalarTypeConfig) *GraphQLScalarType {
-	st := &GraphQLScalarType{}
-	err := invariant(config.Name != "", "Type must be named.")
-	if err != nil {
-		st.err = err
-		return st
-	}
-
-	err = assertValidName(config.Name)
-	if err != nil {
-		st.err = err
-		return st
-	}
-
-	st.Name = config.Name
-	st.Description = config.Description
-
-	err = invariant(
-		config.Serialize != nil,
-		fmt.Sprintf(`%v must provide "serialize" function. If this custom Scalar is `+
-			`also used as an input type, ensure "parseValue" and "parseLiteral" `+
-			`functions are also provided.`, st),
-	)
-	if err != nil {
-		st.err = err
-		return st
-	}
-	if config.ParseValue != nil || config.ParseLiteral != nil {
-		err = invariant(
-			config.ParseValue != nil && config.ParseLiteral != nil,
-			fmt.Sprintf(`%v must provide both "parseValue" and "parseLiteral" functions.`, st),
-		)
-		if err != nil {
-			st.err = err
-			return st
-		}
-	}
-
-	st.scalarConfig = config
-	return st
-}
-
-func (st *GraphQLScalarType) Serialize(value interface{}) interface{} {
-	if st.scalarConfig.Serialize == nil {
-		return value
-	}
-	return st.scalarConfig.Serialize(value)
-}
-func (st *GraphQLScalarType) ParseValue(value interface{}) interface{} {
-	if st.scalarConfig.ParseValue == nil {
-		return value
-	}
-	return st.scalarConfig.ParseValue(value)
-}
-func (st *GraphQLScalarType) ParseLiteral(valueAST ast.Value) interface{} {
-	if st.scalarConfig.ParseLiteral == nil {
-		return nil
-	}
-	return st.scalarConfig.ParseLiteral(valueAST)
-}
-
-func (st *GraphQLScalarType) GetName() string {
-	return st.Name
-}
-func (st *GraphQLScalarType) GetDescription() string {
-	return st.Description
-
-}
-func (st *GraphQLScalarType) String() string {
-	return st.Name
-}
-func (st *GraphQLScalarType) GetError() error {
-	return st.err
-}
-
 func coerceInt(value interface{}) interface{} {
 	switch value := value.(type) {
 	case bool:
